+++ dart/constraint/ContactConstraint.cpp
@@ -48,12 +48,6 @@
 #define DART_CFM 1e-5
 // #define DART_MAX_NUMBER_OF_CONTACTS 32
 
-#define DART_RESTITUTION_COEFF_THRESHOLD 1e-3
-#define DART_FRICTION_COEFF_THRESHOLD 1e-3
-#define DART_BOUNCING_VELOCITY_THRESHOLD 1e-1
-#define DART_MAX_BOUNCING_VELOCITY 1e+2
-#define DART_CONTACT_CONSTRAINT_EPSILON_SQUARED 1e-12
-
 namespace dart {
 namespace constraint {
 
@@ -62,14 +56,22 @@ double ContactConstraint::mErrorReductionParameter = DART_ERP;
 double ContactConstraint::mMaxErrorReductionVelocity = DART_MAX_ERV;
 double ContactConstraint::mConstraintForceMixing = DART_CFM;
 
-constexpr double DART_DEFAULT_FRICTION_COEFF = 1.0;
-constexpr double DART_DEFAULT_RESTITUTION_COEFF = 0.0;
-constexpr double DART_DEFAULT_SLIP_COMPLIANCE = 0.0;
-const Eigen::Vector3d DART_DEFAULT_FRICTION_DIR = Eigen::Vector3d::UnitZ();
-
 //==============================================================================
 ContactConstraint::ContactConstraint(
     collision::Contact& contact, double timeStep)
+  : ContactConstraint(
+      contact,
+      timeStep,
+      DefaultContactSurfaceHandler().createParams(contact, 1u))
+{
+  // Do nothing
+}
+
+//==============================================================================
+ContactConstraint::ContactConstraint(
+    collision::Contact& contact,
+    double timeStep,
+    const ContactSurfaceParams& contactSurfaceParams)
   : ConstraintBase(),
     mTimeStep(timeStep),
     mBodyNodeA(const_cast<dynamics::ShapeFrame*>(
@@ -94,19 +96,10 @@ ContactConstraint::ContactConstraint(
   assert(
       contact.normal.squaredNorm() >= DART_CONTACT_CONSTRAINT_EPSILON_SQUARED);
 
-  const auto* shapeNodeA = const_cast<dynamics::ShapeFrame*>(
-                               contact.collisionObject1->getShapeFrame())
-                               ->asShapeNode();
-  const auto* shapeNodeB = const_cast<dynamics::ShapeFrame*>(
-                               contact.collisionObject2->getShapeFrame())
-                               ->asShapeNode();
-
   //----------------------------------------------
   // Bounce
   //----------------------------------------------
-  const double restitutionCoeffA = computeRestitutionCoefficient(shapeNodeA);
-  const double restitutionCoeffB = computeRestitutionCoefficient(shapeNodeB);
-  mRestitutionCoeff = restitutionCoeffA * restitutionCoeffB;
+  mRestitutionCoeff = contactSurfaceParams.mRestitutionCoeff;
   if (mRestitutionCoeff > DART_RESTITUTION_COEFF_THRESHOLD)
     mIsBounceOn = true;
   else
@@ -115,78 +108,18 @@ ContactConstraint::ContactConstraint(
   //----------------------------------------------
   // Friction
   //----------------------------------------------
-  // TODO(JS): Assume the frictional coefficient can be changed during
-  //           simulation steps.
-  // Update mFrictionCoeff
-  const double primaryFrictionCoeffA
-      = computePrimaryFrictionCoefficient(shapeNodeA);
-  const double primaryFrictionCoeffB
-      = computePrimaryFrictionCoefficient(shapeNodeB);
-  const double secondaryFrictionCoeffA
-      = computeSecondaryFrictionCoefficient(shapeNodeA);
-  const double secondaryFrictionCoeffB
-      = computeSecondaryFrictionCoefficient(shapeNodeB);
-
-  // TODO(JS): Consider providing various ways of the combined friction or
-  // allowing to override this method by a custom method
-  mPrimaryFrictionCoeff
-      = std::min(primaryFrictionCoeffA, primaryFrictionCoeffB);
-  mSecondaryFrictionCoeff
-      = std::min(secondaryFrictionCoeffA, secondaryFrictionCoeffB);
+  mPrimaryFrictionCoeff = contactSurfaceParams.mPrimaryFrictionCoeff;
+  mSecondaryFrictionCoeff = contactSurfaceParams.mSecondaryFrictionCoeff;
   if (mPrimaryFrictionCoeff > DART_FRICTION_COEFF_THRESHOLD
       || mSecondaryFrictionCoeff > DART_FRICTION_COEFF_THRESHOLD)
   {
     mIsFrictionOn = true;
 
-    // Compute slip compliance
-    const double primarySlipComplianceA
-        = computePrimarySlipCompliance(shapeNodeA);
-    const double primarySlipComplianceB
-        = computePrimarySlipCompliance(shapeNodeB);
-    const double secondarySlipComplianceA
-        = computeSecondarySlipCompliance(shapeNodeA);
-    const double secondarySlipComplianceB
-        = computeSecondarySlipCompliance(shapeNodeB);
     // Combine slip compliances through addition
-    mPrimarySlipCompliance = primarySlipComplianceA + primarySlipComplianceB;
-    mSecondarySlipCompliance
-        = secondarySlipComplianceA + secondarySlipComplianceB;
+    mPrimarySlipCompliance = contactSurfaceParams.mPrimarySlipCompliance;
+    mSecondarySlipCompliance = contactSurfaceParams.mSecondarySlipCompliance;
 
-    // Check shapeNodes for valid friction direction unit vectors
-    auto frictionDirA = computeWorldFirstFrictionDir(shapeNodeA);
-    auto frictionDirB = computeWorldFirstFrictionDir(shapeNodeB);
-
-    // check if the friction direction unit vectors have been set
-    bool nonzeroDirA
-        = frictionDirA.squaredNorm() >= DART_CONTACT_CONSTRAINT_EPSILON_SQUARED;
-    bool nonzeroDirB
-        = frictionDirB.squaredNorm() >= DART_CONTACT_CONSTRAINT_EPSILON_SQUARED;
-
-    // only consider custom friction direction if one has nonzero length
-    if (nonzeroDirA || nonzeroDirB)
-    {
-      // if A and B are both set, choose one with smaller friction coefficient
-      // since it's friction properties will dominate
-      if (nonzeroDirA && nonzeroDirB)
-      {
-        if (primaryFrictionCoeffA <= primaryFrictionCoeffB)
-        {
-          mFirstFrictionalDirection = frictionDirA.normalized();
-        }
-        else
-        {
-          mFirstFrictionalDirection = frictionDirB.normalized();
-        }
-      }
-      else if (nonzeroDirA)
-      {
-        mFirstFrictionalDirection = frictionDirA.normalized();
-      }
-      else
-      {
-        mFirstFrictionalDirection = frictionDirB.normalized();
-      }
-    }
+    mFirstFrictionalDirection = contactSurfaceParams.mFirstFrictionalDirection;
 
     // Update frictional direction
     updateFirstFrictionalDirection();
@@ -195,6 +128,9 @@ ContactConstraint::ContactConstraint(
   {
     mIsFrictionOn = false;
   }
+
+  mContactSurfaceMotionVelocity
+      = contactSurfaceParams.mContactSurfaceMotionVelocity;
 
   assert(mBodyNodeA->getSkeleton());
   assert(mBodyNodeB->getSkeleton());
@@ -491,6 +427,9 @@ void ContactConstraint::getInformation(ConstraintInfo* info)
     }
 
     info->b[0] += bouncingVelocity;
+    info->b[0] += mContactSurfaceMotionVelocity.x();
+    info->b[1] += mContactSurfaceMotionVelocity.y();
+    info->b[2] += mContactSurfaceMotionVelocity.z();
 
     // TODO(JS): Initial guess
     // x
@@ -546,6 +485,7 @@ void ContactConstraint::getInformation(ConstraintInfo* info)
     }
 
     info->b[0] += bouncingVelocity;
+    info->b[0] += mContactSurfaceMotionVelocity.x();
 
     // TODO(JS): Initial guess
     // x
@@ -755,163 +695,52 @@ bool ContactConstraint::isActive() const
 double ContactConstraint::computeFrictionCoefficient(
     const dynamics::ShapeNode* shapeNode)
 {
-  assert(shapeNode);
-
-  auto dynamicAspect = shapeNode->getDynamicsAspect();
-
-  if (dynamicAspect == nullptr)
-  {
-    dtwarn << "[ContactConstraint] Attempt to extract friction coefficient "
-           << "from a ShapeNode that doesn't have DynamicAspect. The default "
-           << "value (" << DART_DEFAULT_FRICTION_COEFF << ") will be used "
-           << "instead.\n";
-    return DART_DEFAULT_FRICTION_COEFF;
-  }
-
-  return dynamicAspect->getFrictionCoeff();
+  return DefaultContactSurfaceHandler::computeFrictionCoefficient(shapeNode);
 }
 
 //==============================================================================
 double ContactConstraint::computePrimaryFrictionCoefficient(
     const dynamics::ShapeNode* shapeNode)
 {
-  assert(shapeNode);
-
-  auto dynamicAspect = shapeNode->getDynamicsAspect();
-
-  if (dynamicAspect == nullptr)
-  {
-    dtwarn << "[ContactConstraint] Attempt to extract "
-           << "primary friction coefficient "
-           << "from a ShapeNode that doesn't have DynamicAspect. The default "
-           << "value (" << DART_DEFAULT_FRICTION_COEFF << ") will be used "
-           << "instead.\n";
-    return DART_DEFAULT_FRICTION_COEFF;
-  }
-
-  return dynamicAspect->getPrimaryFrictionCoeff();
+  return DefaultContactSurfaceHandler::computePrimaryFrictionCoefficient(
+      shapeNode);
 }
 
 //==============================================================================
 double ContactConstraint::computeSecondaryFrictionCoefficient(
     const dynamics::ShapeNode* shapeNode)
 {
-  assert(shapeNode);
-
-  auto dynamicAspect = shapeNode->getDynamicsAspect();
-
-  if (dynamicAspect == nullptr)
-  {
-    dtwarn << "[ContactConstraint] Attempt to extract "
-           << "secondary friction coefficient "
-           << "from a ShapeNode that doesn't have DynamicAspect. The default "
-           << "value (" << DART_DEFAULT_FRICTION_COEFF << ") will be used "
-           << "instead.\n";
-    return DART_DEFAULT_FRICTION_COEFF;
-  }
-
-  return dynamicAspect->getSecondaryFrictionCoeff();
+  return DefaultContactSurfaceHandler::computeSecondaryFrictionCoefficient(
+      shapeNode);
 }
 
 //==============================================================================
 double ContactConstraint::computePrimarySlipCompliance(
     const dynamics::ShapeNode* shapeNode)
 {
-  assert(shapeNode);
-
-  auto dynamicAspect = shapeNode->getDynamicsAspect();
-
-  if (dynamicAspect == nullptr)
-  {
-    dtwarn << "[ContactConstraint] Attempt to extract slip compliance "
-           << "from a ShapeNode that doesn't have DynamicAspect. The default "
-           << "value (" << DART_DEFAULT_SLIP_COMPLIANCE << ") will be used "
-           << "instead.\n";
-    return DART_DEFAULT_SLIP_COMPLIANCE;
-  }
-
-  double slipCompliance = dynamicAspect->getPrimarySlipCompliance();
-  if (slipCompliance < 0)
-  {
-    return DART_DEFAULT_SLIP_COMPLIANCE;
-  }
-  return slipCompliance;
+  return DefaultContactSurfaceHandler::computePrimarySlipCompliance(shapeNode);
 }
 
 //==============================================================================
 double ContactConstraint::computeSecondarySlipCompliance(
     const dynamics::ShapeNode* shapeNode)
 {
-  assert(shapeNode);
-
-  auto dynamicAspect = shapeNode->getDynamicsAspect();
-
-  if (dynamicAspect == nullptr)
-  {
-    dtwarn << "[ContactConstraint] Attempt to extract "
-           << "secondary slip compliance "
-           << "from a ShapeNode that doesn't have DynamicAspect. The default "
-           << "value (" << DART_DEFAULT_SLIP_COMPLIANCE << ") will be used "
-           << "instead.\n";
-    return DART_DEFAULT_SLIP_COMPLIANCE;
-  }
-
-  double slipCompliance = dynamicAspect->getSecondarySlipCompliance();
-  if (slipCompliance < 0)
-  {
-    return DART_DEFAULT_SLIP_COMPLIANCE;
-  }
-  return slipCompliance;
+  return DefaultContactSurfaceHandler::computeSecondarySlipCompliance(
+      shapeNode);
 }
 
 //==============================================================================
 Eigen::Vector3d ContactConstraint::computeWorldFirstFrictionDir(
     const dynamics::ShapeNode* shapeNode)
 {
-  assert(shapeNode);
-
-  auto dynamicAspect = shapeNode->getDynamicsAspect();
-
-  if (dynamicAspect == nullptr)
-  {
-    dtwarn << "[ContactConstraint] Attempt to extract friction direction "
-           << "from a ShapeNode that doesn't have DynamicAspect. The default "
-           << "value (" << DART_DEFAULT_FRICTION_DIR.transpose()
-           << ") will be used instead.\n";
-    return DART_DEFAULT_FRICTION_DIR;
-  }
-
-  auto frame = dynamicAspect->getFirstFrictionDirectionFrame();
-  const Eigen::Vector3d& frictionDir
-      = dynamicAspect->getFirstFrictionDirection();
-
-  // rotate using custom frame if it is specified
-  if (frame)
-  {
-    return frame->getWorldTransform().linear() * frictionDir;
-  }
-  // otherwise rotate using shapeNode
-  return shapeNode->getWorldTransform().linear() * frictionDir;
+  return DefaultContactSurfaceHandler::computeWorldFirstFrictionDir(shapeNode);
 }
 
 //==============================================================================
 double ContactConstraint::computeRestitutionCoefficient(
     const dynamics::ShapeNode* shapeNode)
 {
-  assert(shapeNode);
-
-  auto dynamicAspect = shapeNode->getDynamicsAspect();
-
-  if (dynamicAspect == nullptr)
-  {
-    dtwarn << "[ContactConstraint] Attempt to extract restitution coefficient "
-           << "from a ShapeNode that doesn't have DynamicAspect. The default "
-           << "value (" << DART_DEFAULT_RESTITUTION_COEFF << ") will be used "
-           << "instead.\n";
-    return DART_DEFAULT_RESTITUTION_COEFF;
-  }
-
-  return dynamicAspect->getRestitutionCoeff();
+  return DefaultContactSurfaceHandler::computeRestitutionCoefficient(shapeNode);
 }
 
 //==============================================================================