+++ dart/constraint/ContactConstraint.hpp
@@ -35,6 +35,7 @@
 
 #include "dart/collision/CollisionDetector.hpp"
 #include "dart/constraint/ConstraintBase.hpp"
+#include "dart/constraint/ContactSurface.hpp"
 #include "dart/math/MathTypes.hpp"
 
 namespace dart {
@@ -51,6 +52,13 @@ class ContactConstraint : public ConstraintBase
 {
 public:
   /// Constructor
+  ContactConstraint(
+      collision::Contact& contact,
+      double timeStep,
+      const ContactSurfaceParams& contactSurfaceParams);
+
+  /// Constructor
+  DART_DEPRECATED(6.13)
   ContactConstraint(collision::Contact& contact, double timeStep);
 
   /// Destructor
@@ -102,6 +110,7 @@ public:
 
   friend class ConstraintSolver;
   friend class ConstrainedGroup;
+  friend class DefaultContactSurfaceHandler;
 
 protected:
   //----------------------------------------------------------------------------
@@ -138,18 +147,25 @@ protected:
   // Documentation inherited
   bool isActive() const override;
 
+  DART_DEPRECATED(6.13)
   static double computeFrictionCoefficient(
       const dynamics::ShapeNode* shapeNode);
+  DART_DEPRECATED(6.13)
   static double computePrimaryFrictionCoefficient(
       const dynamics::ShapeNode* shapeNode);
+  DART_DEPRECATED(6.13)
   static double computeSecondaryFrictionCoefficient(
       const dynamics::ShapeNode* shapeNode);
+  DART_DEPRECATED(6.13)
   static double computePrimarySlipCompliance(
       const dynamics::ShapeNode* shapeNode);
+  DART_DEPRECATED(6.13)
   static double computeSecondarySlipCompliance(
       const dynamics::ShapeNode* shapeNode);
+  DART_DEPRECATED(6.13)
   static Eigen::Vector3d computeWorldFirstFrictionDir(
       const dynamics::ShapeNode* shapenode);
+  DART_DEPRECATED(6.13)
   static double computeRestitutionCoefficient(
       const dynamics::ShapeNode* shapeNode);
 
@@ -206,7 +222,7 @@ private:
   /// Primary Coefficient of Friction
   double mPrimaryFrictionCoeff;
 
-  /// Primary Coefficient of Friction
+  /// Secondary Coefficient of Friction
   double mSecondaryFrictionCoeff;
 
   /// Primary Coefficient of Slip Compliance
@@ -217,6 +233,12 @@ private:
 
   /// Coefficient of restitution
   double mRestitutionCoeff;
+
+  /// Velocity of the contact independent of friction
+  /// x = vel. in direction of contact normal
+  /// y = vel. in first friction direction
+  /// z = vel. in second friction direction
+  Eigen::Vector3d mContactSurfaceMotionVelocity;
 
   /// Whether this contact is self-collision.
   bool mIsSelfCollision;